Revert "Refactor: queries from the schema file"
This reverts commit 7bf22e69dad7cac97741d7348b16da3f5c34d498.
