+++ src/components/sidebar/sidebar.tsx
@@ -2,9 +2,14 @@ import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import React from "react";
 
 import { cookies } from "next/headers";
-import { getCollaboratingWorkspaces, getFolders, getPrivateWorkspaces, getSharedWorkspaces, getUserSubscriptionStatus } from "@/lib/supabase/queries";
+import { getFolders, getUserSubscriptionStatus } from "@/lib/supabase/queries";
 import { redirect } from "next/navigation";
 import { twMerge } from "tailwind-merge";
+import {
+  getCollaboratingWorkspaces,
+  getPrivateWorkspaces,
+  getSharedWorkspaces,
+} from "@/lib/supabase/schema";
 import WorkspaceDropdown from "./workspace-dropdown";
 
 interface SidebarProps {
