+++ src/lib/supabase/queries.ts
@@ -1,8 +1,8 @@
 "use server";
-import { and, eq, ilike, notExists } from "drizzle-orm";
-import { collaborators, folders, users, workspaces } from "../../../migrations/schema";
+import { eq } from "drizzle-orm";
+import { folders, workspaces } from "../../../migrations/schema";
 import db from "./db";
-import { Folder, Subscription, User, workspace } from "./supabase.types";
+import { Folder, Subscription, workspace } from "./supabase.types";
 import { validate } from 'uuid';
 
 export const createWorkspace = async (workspace: workspace) => {
@@ -46,95 +46,4 @@ export const getFolders = async (workspaceId: string) => {
   } catch (error) {
     return { data: null, error: 'Error' };
   }
-};
-
-export const getPrivateWorkspaces = async (userId: string) => {
-  if (!userId) return [];
-  const privateWorkspaces = (await db
-    .select({
-      id: workspaces.id,
-      createdAt: workspaces.createdAt,
-      workspaceOwner: workspaces.workspaceOwner,
-      title: workspaces.title,
-      iconId: workspaces.iconId,
-      data: workspaces.data,
-      inTrash: workspaces.inTrash,
-      logo: workspaces.logo,
-      bannerUrl: workspaces.bannerUrl,
-    })
-    .from(workspaces)
-    .where(
-      and(
-        notExists(
-          db
-            .select()
-            .from(collaborators)
-            .where(eq(collaborators.workspaceId, workspaces.id))
-        ),
-        eq(workspaces.workspaceOwner, userId)
-      )
-    )) as workspace[];
-  return privateWorkspaces;
-};
-
-export const getCollaboratingWorkspaces = async (userId: string) => {
-  if (!userId) return [];
-  const collaboratedWorkspaces = (await db
-    .select({
-      id: workspaces.id,
-      createdAt: workspaces.createdAt,
-      workspaceOwner: workspaces.workspaceOwner,
-      title: workspaces.title,
-      iconId: workspaces.iconId,
-      data: workspaces.data,
-      inTrash: workspaces.inTrash,
-      logo: workspaces.logo,
-      bannerUrl: workspaces.bannerUrl,
-    })
-    .from(users)
-    .innerJoin(collaborators, eq(users.id, collaborators.userId))
-    .innerJoin(workspaces, eq(collaborators.workspaceId, workspaces.id))
-    .where(eq(users.id, userId))) as workspace[];
-  return collaboratedWorkspaces;
-};
-
-export const getSharedWorkspaces = async (userId: string) => {
-  if (!userId) return [];
-  const sharedWorkspaces = (await db
-    .selectDistinct({
-      id: workspaces.id,
-      createdAt: workspaces.createdAt,
-      workspaceOwner: workspaces.workspaceOwner,
-      title: workspaces.title,
-      iconId: workspaces.iconId,
-      data: workspaces.data,
-      inTrash: workspaces.inTrash,
-      logo: workspaces.logo,
-      bannerUrl: workspaces.bannerUrl,
-    })
-    .from(workspaces)
-    .orderBy(workspaces.createdAt)
-    .innerJoin(collaborators, eq(workspaces.id, collaborators.workspaceId))
-    .where(eq(workspaces.workspaceOwner, userId))) as workspace[];
-  return sharedWorkspaces;
-};
-
-export const addCollaborators = async (users: User[], workspaceId: string) => {
-  const response = users.forEach(async (user: User) => {
-    const userExists = await db.query.collaborators.findFirst({
-      where: (u, { eq }) =>
-        and(eq(u.userId, user.id), eq(u.workspaceId, workspaceId)),
-    });
-    if (!userExists)
-      await db.insert(collaborators).values({ workspaceId, userId: user.id });
-  });
-};
-
-export const getUsersFromSearch = async (email: string) => {
-  if (!email) return [];
-  const accounts = db
-    .select()
-    .from(users)
-    .where(ilike(users.email, `${email}%`));
-  return accounts;
 };
